Spring Security

+++ Sales-System/src/main/java/com/example/Sales/System/service/ProductService.java
@@ -6,6 +6,7 @@ import com.example.Sales.System.entity.Category;
 import com.example.Sales.System.entity.Product;
 import com.example.Sales.System.entity.User;
 import com.example.Sales.System.enums.Role;
+import com.example.Sales.System.errors.RestrictedAccess;
 import com.example.Sales.System.errors.WrongUserType;
 import com.example.Sales.System.mapper.Mapper;
 import com.example.Sales.System.repository.CategoryRepository;
@@ -13,10 +14,13 @@ import com.example.Sales.System.repository.ProductRepository;
 import com.example.Sales.System.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -26,16 +30,33 @@ public class ProductService {
     private final UserRepository userRepository;
     private final CategoryRepository categoryRepository;
 
+    private User getAuthenticatedUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null && authentication.getPrincipal() instanceof User) {
+            return (User) authentication.getPrincipal();
+        }
+        throw new IllegalStateException("User not authenticated");
+    }
+
     public List<ProductDTO> getAllProducts() {
-        return mapper.productListToProductDTOList(productRepository.findAll());
+        User user = getAuthenticatedUser();
+        if (user.getRole() == Role.ADMIN)
+            return mapper.productListToProductDTOList(productRepository.findAll());
+        else if (user.getRole() == Role.SELLER || user.getRole() == Role.BOTH)
+            return mapper.productListToProductDTOList(productRepository.findBySeller(user));
+        else
+            throw new RestrictedAccess("Restricted Access");
     }
 
     public void createProduct(ProductDTO productDTO) {
+        User sender = getAuthenticatedUser();
+        if(!Objects.equals(sender.getId(), productDTO.getSeller().getId()))
+            throw new RestrictedAccess("Restricted Access");
         Product product = mapper.productDTOToProduct(productDTO);
-        User user = userRepository.findById(productDTO.getSeller().getId()).orElseThrow();
-        if (user.getRole()!= Role.SELLER &&user.getRole()!= Role.BOTH)
+        User seller = userRepository.findById(productDTO.getSeller().getId()).orElseThrow();
+        if (seller.getRole()!= Role.SELLER &&seller.getRole()!= Role.BOTH)
             throw new WrongUserType("not a seller");
-        product.setSeller(user);
+        product.setSeller(seller);
         HashSet<Category> categories = new HashSet<>();
         for (CategoryDTO categoryDTO : productDTO.getCategory()) {
             Category category = categoryRepository.findByName(categoryDTO.getName());
@@ -53,6 +74,9 @@ public class ProductService {
     @Transactional
     public void updateProduct(ProductDTO productDTO) {
         Product product = productRepository.findById(productDTO.getId()).orElseThrow();
+        User sender = getAuthenticatedUser();
+        if (!Objects.equals(product.getSeller().getId(), sender.getId()))
+            throw new RestrictedAccess("Restricted Access");
         if (productDTO.getName() != null)
             product.setName(productDTO.getName());
         if (productDTO.getPrice() != null)
@@ -84,6 +108,10 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
+        User sender = getAuthenticatedUser();
+        Product product=productRepository.findById(id).orElseThrow();
+        if (!Objects.equals(product.getSeller().getId(), sender.getId()))
+            throw new RestrictedAccess("Restricted Access");
         productRepository.deleteById(id);
     }
